Cast sigma coordinate and area tensors to float at init (#658)

+++ fme/fme/core/data_loading/_xarray.py
@@ -119,8 +119,8 @@ def get_sigma_coordinates(ds: xr.Dataset) -> SigmaCoordinates:
         )
 
     return SigmaCoordinates(
-        ak=torch.as_tensor(ak_list, device=fme.get_device()),
-        bk=torch.as_tensor(bk_list, device=fme.get_device()),
+        ak=torch.as_tensor(ak_list, device=fme.get_device(), dtype=torch.float),
+        bk=torch.as_tensor(bk_list, device=fme.get_device(), dtype=torch.float),
     )
